requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
ngrok==0.8.1
openai-whisper==20230314
